requirements.txt
torch
transformers
Pillow
faiss-cpu
PyMuPDF
pdfplumber
streamlit
torch==2.0.0
transformers==4.30.0
Pillow==9.3.0
streamlit==1.24.0
pinecone-client==2.2.1
PyMuPDF==1.22.1

